small change in ActionDispatcher.cpp, is faster, but if Action was holding body's id it is possible that it can be faster
result: 107 seconds is faster compared to 111 seconds

tere are a lot of places where we can do improvement like this. if we do code change like here - in every possible places - yade can gain a lot of speed.

basically I've discovered that every temporary variable of type shared_ptr<> costs 3 seconds, so we should avoid creating them and instead pass them around by references - whenever possible. 

And very iportant to note is that this speed improvement is possible ONLY because we have introduced Action, which is more generic! without this class it would be impossible to get faster. So it means that good generic code allows faster code - at least sometimes - as we have example if this here. The same was with containers - introducing containers made code slower, but it made possible to make improvements in other places (because containers increased genericity) and as a result the code got faster.

M    yade/src/yade/Engine/ActionDispatcher.cpp
M    yade/projects/kdevelop/yade.kdevelop