Some cleaning, involving changing variable names, like Body::interactingGeometry. Which caused almost all the files to be modified.
And most imporatntly added a primitive example of doing calculations with
tetrahedrons. It is for educational purposes about how to add new stuff to
yade. Later I will have to come back to this commit, and extract from it a
useful example for the manual.

This example is for Václav Šmilauer, because he wants to write a complete
simulation with tetrahedrons. So this is a starting point.