Added drawMask to OpenGLRenderingEngine, it is now possible to see only bodies that have a certain groupMask
Added Quadrilateral, but it is in fact a hack, that makes it possible to see
the evolution of strain field. It is much simpler to add something as a
component of OpenGLRenderingEngine (the drawing loop) than as a component of
engines (the calculation loop). This fact is a string indication that design of
OpenGLRenderingEngine is much better than design of Metabody->engines. And so
engines must be eliminated to simplify their use. Whole process will get automated.

So currently we have a Quadrilateral that is not really a Quadrilateral, but
just a quick hack to see evolution of strain field. The real solution will be
to change the data structure. And this solution is coming ;) (ut don't know when)

Anyway, adding it, gave me a lot of thoughts about the current design of data.
I won't leave this unfixed, this would be too bad :>