1. Main commit: import a geometry of walls from a STL file.
An STL file describes a raw unstructured triangulated surface by the unit
normal and vertices (ordered by the right-hand rule) of the triangles using a
three-dimensional Cartesian coordinate system (see
http://en.wikipedia.org/wiki/STL_%28file_format%29).  Allows binary and ascii
formats of a STL file.

The STLImporterTest preprocessor provided a example. In order to work it need
hourglass.stl file located in YADESOURCE/example.

(For creation a STL file I use blender, see http://www.blender.org)

The STL file is imported into YADE as follows: the vertices, edges and facets
are distinct bodies; the facets which have a common edge and lies on a plane
are joined, herewith maximum number of vertices in the joined facet are limited
(by default 4).

A facet is a convex polygon and describes by N vertices (3,4 or more). The
origin vertex of a facet lies in the facet se3 position.  The others vertices
is describes relative to the origin vertex.  Let the vector n_f is the facets
normal.  Then the distance from the centre x of a sphere to a facets plane will be
equal to the module of the vector d, 
d = n_f.Dot(x-x'),
where vector x' is a any vertex of the facet (for example, the origin vertex).
Provided that the projection of the centre of a sphere to facets plane lies in
side perimeter, i.e.
(x'_j+1 - x'_j).Dot(x - d - x'_j)>0, j=1,...,N-1.

A edge is a vector with origin in the edge se3 position.
Two normals n1, n2 are connected with an edge. This normals being normals to the planes
passing through an edge perpendicular to two facets which the given edge divides 
(an edge cannot divide more than two facets).
Contact between sphere and an edge is possible only if the sphere centre is
simultaneously on a positive side of both planes, i.e.
n_j.Dot(x-x'_j)>=0, j=1,2
where x'_j is a point lies on the edge j.
Let x'_0 is the edge origin, l is the length of the edge, n'_0 is the unit
direction vector. Then the distance from a point x to the edge (_not_
including its final points) will be equal to the vector module d,
d = x - x'_0 - cn'_0,
if c = n'_0.Dot(x-x'_0)/l^2 satisfies to a condition 0<c<1.

A vertex is a point lies in the vertex se3 position. So much normals are
connected with vertex, how many edges converge in the given vertex. I.e.
normals are normals to the planes passing through vertex and perpendicular
edges converging in vertex. Contact of sphere to vertex is possible, if the
sphere centre is on a positive side of all planes (by analogy to edges).

2. Add new collider: SpatialQuickSortCollider is analog PersistentSAPCollider,
but seems more fast.

3. SConstruct been modified in order to works with symbol links in the YADESOURCE directory.