NewtonsDampedLaw : fixing a small mistake and cleaning the code. IMPORTANT NOTE : Forces/Moments are not affected by damping now, only accelerations are modified. The result is the same in terms of damping, but users can now get the exact values of forces/moments on bodies at any point during a timestep.
TriaxialCompressionEngine : compute unbalanced force even in LIMBO state

GlobalStiffnessTimeStepper : default safety coefficient is 1

LeapFrogIntegrators : remove previous velocities and  make this 2nd
order scheme.

IMPORTANT NOTE (2) : the damping obtained with 
NewtonsDampedLaw is not 
exactly the
same as the one in CundallNonViscous_X_Damping. Just because it is 
too much work to implement the same algorithm in splitted classes 
(it needs 
new includes and libs for some classes, and it also needs reordering 
integrating engines in all existing preprocessors).
I will no longer maintain those 6 integrating classes, as it seems 
useless and painfull to me.